Refactor `SignInPage`
