+++ src/pages/sign-in/sign-in.html
@@ -7,16 +7,14 @@
     <form [formGroup]="form" (ngSubmit)="signIn()">
       <ion-item>
         <ion-label floating>Username</ion-label>
-        <ion-input formControlName="username"
-          id="username"
-          type="text">
+        <ion-input formControlName="username" id="username" type="text">
         </ion-input>
         <i class="icon ion-alert-circled error"></i>
       </ion-item>
 
       <div *ngIf="username.touched && username.invalid" class="form-error-message">
         <div *ngIf="username.errors.required">
-            Username is required.
+          Username is required.
         </div>
         <div *ngIf="username.errors.minlength">
           Username should be minimum {{ username.errors.minlength.requiredLength }} characters.
@@ -25,10 +23,7 @@
 
       <ion-item class="input-item">
         <ion-label floating>Password</ion-label>
-        <ion-input
-          formControlName="password"
-          id="password"
-          type="password">
+        <ion-input formControlName="password" id="password" type="password">
         </ion-input>
       </ion-item>
