+++ src/pages/sign-in/sign-in.ts
@@ -1,11 +1,11 @@
 import { Component } from '@angular/core';
 import { FormBuilder, FormGroup, Validators } from '@angular/forms';
-import { NavController, NavParams } from 'ionic-angular';
+import { NavController } from 'ionic-angular';
 
 import { HomePage } from '../home/home';
 
-import { CommonProvider } from './../../providers/common/common';
-import { AuthProvider } from './../../providers/auth/auth';
+import { CommonProvider } from '../../providers/common/common';
+import { AuthProvider } from '../../providers/auth/auth';
 
 @Component({
   selector: 'page-sign-in',
@@ -14,13 +14,13 @@ import { AuthProvider } from './../../providers/auth/auth';
 export class SignInPage {
   public form: FormGroup;
 
-  constructor(public navCtrl: NavController,
-    public navParams: NavParams,
-    private formBuilder: FormBuilder,
-    private commonProvider: CommonProvider,
-    private authProvider: AuthProvider) {
-
-    this.form = this.formBuilder.group({
+  constructor(
+    private _formBuilder: FormBuilder,
+    private _navCtrl: NavController,
+    private _commonProvider: CommonProvider,
+    private _authProvider: AuthProvider
+  ) {
+    this.form = this._formBuilder.group({
       username: ['', [Validators.required, Validators.minLength(4)]],
       password: ['', [Validators.required, Validators.minLength(4)]],
     });
@@ -37,15 +37,17 @@ export class SignInPage {
   public signIn() {
     const username: string = this.form.value.username;
     const password: string = this.form.value.password;
-    this.authProvider.signIn(username, password).subscribe(
+    this._authProvider.signIn(username, password).subscribe(
       data => {
-        this.commonProvider.setAuthenticationData(data);
-        this.navCtrl.setRoot(HomePage);
+        this._commonProvider.setAuthenticationData(data);
+        this._navCtrl.setRoot(HomePage);
       },
       error => {
         for (const key in error.error) {
           if (this.form.get(key)) {
-            this.form.get(key).setErrors({remote: error.error[key]});
+            this.form.get(key).setErrors({
+              remote: error.error[key]
+            });
           }
         }
       }
